extract the cluster initialization logic from L2PQQuantizer, move symmetricDistance(...) method from L2PQQuantizer to AbstractQuantizer
