+++ vectoriadb-index/src/main/java/jetbrains/vectoriadb/index/ClusterInitializer.java
@@ -1,0 +1,18 @@
+package jetbrains.vectoriadb.index;
+
+import org.jetbrains.annotations.NotNull;
+
+import java.lang.foreign.MemorySegment;
+
+public interface ClusterInitializer {
+    void initializeCentroids(
+            MemorySegment pqVectors,
+            long numVectors,
+            float[] distanceTable,
+            int quantizersCount,
+            int codeBaseSize,
+            byte[] pqCentroids,
+            int numClusters,
+            @NotNull ProgressTracker progressTracker
+    );
+}
