+++ vectoriadb-index/src/main/java/jetbrains/vectoriadb/index/L2PQQuantizer.java
@@ -16,7 +16,6 @@
 package jetbrains.vectoriadb.index;
 
 import it.unimi.dsi.fastutil.ints.IntArrayList;
-import org.apache.commons.rng.simple.RandomSource;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.DataInputStream;
@@ -40,6 +39,8 @@ class L2PQQuantizer extends AbstractQuantizer {
     private int vectorsCount;
     private final Arena arena;
     private MemorySegment pqVectors;
+    @NotNull
+    private ClusterInitializer clusterInitializer;
 
     //1st dimension quantizer index
     //2nd index of code inside code book
@@ -48,8 +49,17 @@ class L2PQQuantizer extends AbstractQuantizer {
 
     L2PQQuantizer() {
         this.arena = Arena.openShared();
+        this.clusterInitializer = new KMeansPlusPlusClusterInitializer();
     }
 
+    L2PQQuantizer(@NotNull ClusterInitializer clusterInitializer) {
+        this.arena = Arena.openShared();
+        this.clusterInitializer = clusterInitializer;
+    }
+
+    public void setClusterInitializer(@NotNull ClusterInitializer clusterInitializer) {
+        this.clusterInitializer = clusterInitializer;
+    }
 
     @Override
     public void generatePQCodes(int vectorsDimension, int compressionRatio, VectorReader vectorReader,
@@ -199,15 +209,14 @@ class L2PQQuantizer extends AbstractQuantizer {
     }
 
     @Override
-    public IntArrayList[] splitVectorsByPartitions(int numClusters, int iterations, DistanceFunction distanceFunction,
-                                                   boolean initializeCentroidsWithKMeansPlusPlus, @NotNull ProgressTracker progressTracker) {
+    public IntArrayList[] splitVectorsByPartitions(int numClusters, int iterations, DistanceFunction distanceFunction, @NotNull ProgressTracker progressTracker) {
         progressTracker.pushPhase("split vectors by partitions",
                 "partitions count", String.valueOf(numClusters));
         try {
             var distanceTables = buildDistanceTables(centroids, quantizersCount, subVectorSize, distanceFunction);
             var pqCentroids = new byte[numClusters * quantizersCount];
 
-            calculateClusters(numClusters, iterations, vectorsCount, pqCentroids, distanceTables, initializeCentroidsWithKMeansPlusPlus, progressTracker);
+            calculateClusters(numClusters, iterations, vectorsCount, pqCentroids, distanceTables, progressTracker);
 
             var vectorsByPartitions = new IntArrayList[numClusters];
             for (int i = 0; i < numClusters; i++) {
@@ -238,14 +247,14 @@ class L2PQQuantizer extends AbstractQuantizer {
     }
 
     @Override
-    public float[][] calculateCentroids(int clustersCount, int iterations, DistanceFunction distanceFunction, boolean initializeCentroidsWithKMeansPlusPlus, @NotNull ProgressTracker progressTracker) {
+    public float[][] calculateCentroids(int clustersCount, int iterations, DistanceFunction distanceFunction, @NotNull ProgressTracker progressTracker) {
         progressTracker.pushPhase("calculate centroids");
         try {
             var numVectors = (int) (pqVectors.byteSize() / quantizersCount);
             var distanceTables = buildDistanceTables(centroids, quantizersCount, subVectorSize, distanceFunction);
             var pqCentroids = new byte[clustersCount * quantizersCount];
 
-            calculateClusters(clustersCount, iterations, numVectors, pqCentroids, distanceTables, initializeCentroidsWithKMeansPlusPlus, progressTracker);
+            calculateClusters(clustersCount, iterations, numVectors, pqCentroids, distanceTables, progressTracker);
 
             var result = new float[clustersCount][];
             for (int i = 0; i < clustersCount; i++) {
@@ -260,7 +269,7 @@ class L2PQQuantizer extends AbstractQuantizer {
 
     private void calculateClusters(int numClusters, int iterations, long numVectors,
                                    byte[] pqCentroids, float[] distanceTables,
-                                   boolean initializeCentroidsWithKMeansPlusPlus, @NotNull ProgressTracker progressTracker) {
+                                   @NotNull ProgressTracker progressTracker) {
         var cores = (int) Math.min(Runtime.getRuntime().availableProcessors(), numVectors);
         progressTracker.pushPhase("PQ k-means clustering",
                 "max iterations", String.valueOf(iterations),
@@ -270,11 +279,7 @@ class L2PQQuantizer extends AbstractQuantizer {
         try (var arena = Arena.openShared()) {
             var centroidIndexes = arena.allocate(numVectors * Integer.BYTES, ValueLayout.JAVA_INT.byteAlignment());
 
-            if (initializeCentroidsWithKMeansPlusPlus) {
-                initializeCentroidsKMeansPlusPlus(pqVectors, numVectors, quantizersCount, distanceTables, numClusters, pqCentroids, progressTracker);
-            } else {
-                initializeCentroidsRandomly(pqVectors, numVectors, quantizersCount, numClusters, pqCentroids);
-            }
+            clusterInitializer.initializeCentroids(pqVectors, numVectors, distanceTables, quantizersCount, codeBaseSize, pqCentroids, numClusters, progressTracker);
 
             var histogram = new float[numClusters * quantizersCount * codeBaseSize];
             var atomicIntegerHistogram = new int[numClusters * quantizersCount * codeBaseSize];
@@ -390,116 +395,6 @@ class L2PQQuantizer extends AbstractQuantizer {
         }
     }
 
-    private void initializeCentroidsRandomly(MemorySegment pqVectors, long numVectors, int quantizersCount, int numClusters, byte[] pqCentroids) {
-        var rng = RandomSource.XO_RO_SHI_RO_128_PP.create();
-        for (int i = 0; i < numClusters; i++) {
-            var vecIndex = rng.nextLong(numVectors);
-            MemorySegment.copy(pqVectors, ValueLayout.JAVA_BYTE, vecIndex * quantizersCount, pqCentroids, i * quantizersCount, quantizersCount);
-        }
-    }
-
-    /**
-     * A smarter than random centroid initialization technique.
-     * 1. Initialize the first centroid randomly.
-     * 2. Calculate the distance to the closest already chosen centroid `distanceToClosestCentroid` for all the vectors.
-     * 3. Choose a vector to become a new centroid based on the weighted probability proportional to `distanceToClosestCentroid^2`.
-     * 4. Repeat from 2 until all the centroid initialized.
-     * */
-    public void initializeCentroidsKMeansPlusPlus(MemorySegment pqVectors, long numVectors, int quantizersCount, float[] distanceTable, int numClusters, byte[] pqCentroids, @NotNull ProgressTracker progressTracker) {
-        progressTracker.pushPhase("Centroids initialization");
-        try (var arena = Arena.openShared()) {
-            var rng = RandomSource.XO_RO_SHI_RO_128_PP.create();
-
-            // initialize the first centroid randomly
-            var firstCentroidIndex = rng.nextLong(numVectors);
-            MemorySegment.copy(pqVectors, ValueLayout.JAVA_BYTE, firstCentroidIndex * quantizersCount, pqCentroids, 0, quantizersCount);
-
-            // here we keep distance to the closest centroid for every vector
-            var distancesToClosestCentroid = arena.allocate(numVectors * Float.BYTES, ValueLayout.JAVA_FLOAT.byteAlignment());
-            // we do not know vectors' distances to any centroids, so make them infinite
-            for (long vectorIdx = 0; vectorIdx < numVectors; vectorIdx++) {
-                distancesToClosestCentroid.setAtIndex(ValueLayout.JAVA_FLOAT, vectorIdx, Float.MAX_VALUE);
-            }
-
-            var cores = (int) Math.min(Runtime.getRuntime().availableProcessors(), numVectors);
-            try (var executors = Executors.newFixedThreadPool(cores, r -> {
-                var thread = new Thread(r);
-                thread.setName("pq-kmeans-cluster-initialization-" + thread.threadId());
-                return thread;
-            })) {
-                var assignmentSize = (numVectors + cores - 1) / cores;
-                var futures = new Future[cores];
-                var sumSquaredDistancesToClosestCentroid = new double[cores];
-
-                for (int newCentroidIdx = 1; newCentroidIdx < numClusters; newCentroidIdx++) {
-                    progressTracker.pushPhase("Initialize centroid " + newCentroidIdx);
-                    try {
-                        var lastCentroidIdx = newCentroidIdx - 1;
-                        var mtProgressTracker = new BoundedMTProgressTrackerFactory(cores, progressTracker);
-                        // map
-                        for (int i = 0; i < cores; i++) {
-                            var start = i * assignmentSize;
-                            var end = Math.min(start + assignmentSize, numVectors);
-                            var id = i;
-
-                            futures[i] = executors.submit(() -> {
-                                try (var localTracker = mtProgressTracker.createThreadLocalTracker(id)) {
-                                    var localSize = end - start;
-                                    double sumSquaredDistanceToClosestCentroid = 0f;
-                                    for (long k = 0; k < localSize; k++) {
-                                        long vectorIdx = start + k;
-                                        float distanceToClosestCentroid = distancesToClosestCentroid.getAtIndex(ValueLayout.JAVA_FLOAT, vectorIdx);
-                                        // process the distance to the last initialized centroid
-                                        float distanceToLastCentroid = symmetricDistance(pqVectors, vectorIdx, pqCentroids, lastCentroidIdx, distanceTable, quantizersCount, codeBaseSize);
-                                        if (distanceToLastCentroid < distanceToClosestCentroid) {
-                                            distanceToClosestCentroid = distanceToLastCentroid;
-                                            distancesToClosestCentroid.setAtIndex(ValueLayout.JAVA_FLOAT, vectorIdx, distanceToClosestCentroid);
-                                        }
-                                        sumSquaredDistanceToClosestCentroid += (distanceToClosestCentroid * distanceToClosestCentroid);
-                                        localTracker.progress(k * 100.0 / localSize);
-                                    }
-                                    sumSquaredDistancesToClosestCentroid[id] = sumSquaredDistanceToClosestCentroid;
-                                }
-                            });
-                        }
-
-                        for (var future : futures) {
-                            try {
-                                future.get();
-                            } catch (InterruptedException | ExecutionException e) {
-                                throw new RuntimeException("Error during cluster initialization phase in PQ kmeans clustering.", e);
-                            }
-                        }
-
-                        // reduce
-                        double sumSquaredDistanceToClosestCentroid = 0;
-                        for (var i = 0; i < cores; i++) {
-                            sumSquaredDistanceToClosestCentroid += sumSquaredDistancesToClosestCentroid[i];
-                        }
-
-                        // choose a vector to become a new centroid
-                        var randomSum = rng.nextDouble(sumSquaredDistanceToClosestCentroid);
-                        double runningSum = 0;
-                        long vectorToBecomeCentroidIdx = 0;
-                        while (vectorToBecomeCentroidIdx < numVectors - 1) {
-                            var distance = distancesToClosestCentroid.getAtIndex(ValueLayout.JAVA_FLOAT, vectorToBecomeCentroidIdx);
-                            runningSum += (distance * distance);
-                            if (runningSum >= randomSum) {
-                                break;
-                            }
-                            vectorToBecomeCentroidIdx++;
-                        }
-                        MemorySegment.copy(pqVectors, ValueLayout.JAVA_BYTE, vectorToBecomeCentroidIdx * quantizersCount, pqCentroids, newCentroidIdx * quantizersCount, quantizersCount);
-                    } finally {
-                        progressTracker.pullPhase(); // Another centroid initialized
-                    }
-                }
-            }
-        } finally {
-            progressTracker.pullPhase(); // Initialization completed
-        }
-    }
-
     static void generateHistogram(final MemorySegment pqVectors,
                                   final MemorySegment clusters,
                                   final int quantizersCount,
@@ -570,8 +465,7 @@ class L2PQQuantizer extends AbstractQuantizer {
         var vectorsCount = centroids.length / quantizersCount;
 
         for (int index = 0; index < vectorsCount; index++) {
-            var distance = symmetricDistance(pqVectors, vectorIndex, centroids, index, distanceTable,
-                    quantizersCount, codeBaseSize);
+            var distance = symmetricDistance(pqVectors, vectorIndex, centroids, index, distanceTable, quantizersCount, codeBaseSize);
 
             if (distance < minDistance) {
                 minDistance = distance;
@@ -593,8 +487,7 @@ class L2PQQuantizer extends AbstractQuantizer {
         var vectorsCount = centroids.length / quantizersCount;
 
         for (int index = 0; index < vectorsCount; index++) {
-            var distance = symmetricDistance(pqVectors, vectorIndex, centroids, index, distanceTable,
-                    quantizersCount, codeBaseSize);
+            var distance = symmetricDistance(pqVectors, vectorIndex, centroids, index, distanceTable, quantizersCount, codeBaseSize);
 
             if (distance < firstMinDistance) {
                 firstMinDistance = distance;
@@ -777,26 +670,6 @@ class L2PQQuantizer extends AbstractQuantizer {
                     result[firstIndex] = result[secondIndex];
                 }
             }
-        }
-
-        return result;
-    }
-
-
-    static float symmetricDistance(MemorySegment pqVectors, long storedEncodedVectorIndex, byte[] encodedVectors, int encodedVectorIndex,
-                                   float[] distanceTables, int quantizersCount, int codeBaseSize) {
-        float result = 0.0f;
-
-        var firstPqBase = MatrixOperations.twoDMatrixIndex(quantizersCount, storedEncodedVectorIndex, 0);
-        var secondPqBase = MatrixOperations.twoDMatrixIndex(quantizersCount, encodedVectorIndex, 0);
-
-        for (int i = 0; i < quantizersCount; i++) {
-            var firstPqCode = Byte.toUnsignedInt(pqVectors.get(ValueLayout.JAVA_BYTE, firstPqBase + i));
-            var secondPwCode = Byte.toUnsignedInt(encodedVectors[secondPqBase + i]);
-
-            var distanceIndex = MatrixOperations.threeDMatrixIndex(codeBaseSize, codeBaseSize,
-                    i, firstPqCode, secondPwCode);
-            result += distanceTables[distanceIndex];
         }
 
         return result;
